.history/requirements_20251221181225.txt
python-dotenv
mlflow
pandas
numpy
scikit-learn
pymango

